requirements.txt
flask
pandas
openpyxl
fpdf2
jinja2
matplotlib
gunicorn
werkzeug
